requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
Pillow
django-cors-headers
cloudinary
reportlab==4.2.0
openpyxl==3.1.4
itsdangerous
Werkzeug==3.0.3
qrcode
requests
cloudinary
psycopg2-binary
pymupdf
